move extensions to a standlone file
